Move MetalEnvMockCCL tests into tt_metal context integration suite.
Relocate mock+fabric CCL regression tests from unit_tests_ttnn into test_integration.cpp per review, linking TTNN into unit_tests_context.
