+++ tests/ttnn/unit_tests/gtests/sources.cmake
@@ -30,7 +30,6 @@ set(UNIT_TESTS_TTNN_BASIC_SOURCES
     test_matmul.cpp
     test_matmul_multicore.cpp
     test_matmul_sweep.cpp
-    test_metal_env_mock_ccl_repro.cpp
     test_reduction.cpp
     test_relational_int.cpp
     test_rsub_int.cpp
